Rename deploy to singletons.
