+++ src/singletons.js
